# frontend/openapi.json
{
  "components": {
    "schemas": {
...
      "ConfigDetail": {
        "properties": {
          "date_col": {
            "title": "Date Col",
            "type": "string"
          },
          "granularities": {
            "items": {
              "enum": [
                "day",
                "week",
                "month"
              ],
              "type": "string"
            },
            "title": "Granularities",
            "type": "array"
          },
          "group_col": {
            "title": "Group Col",
            "type": "string"
          },
          "groups": {
            "items": {
              "$ref": "#/components/schemas/MetricGroup"
            },
            "title": "Groups",
            "type": "array"
          },
          "id": {
            "title": "Id",
            "type": "string"
          },
          "lifecycle_col": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Lifecycle Col"
          },
          "range_group_col": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Range Group Col"
          },
          "range_group_defaults": {
            "default": [],
            "items": {
              "$ref": "#/components/schemas/RangeGroup"
            },
            "title": "Range Group Defaults",
            "type": "array"
          },
          "range_group_name": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Range Group Name"
          },
          "range_group_values": {
            "default": [],
            "items": {
              "type": "number"
            },
            "title": "Range Group Values",
            "type": "array"
          },
          "tabs": {
            "items": {
              "type": "string"
            },
            "title": "Tabs",
            "type": "array"
          },
          "title": {
            "title": "Title",
            "type": "string"
          },
          "user_group_cols": {
            "items": {
              "type": "string"
            },
            "title": "User Group Cols",
            "type": "array"
          }
        },
        "required": [
          "id",
          "title",
          "date_col",
          "group_col",
          "user_group_cols",
          "granularities",
          "groups",
          "tabs"
        ],
        "title": "ConfigDetail",
        "type": "object"
      },
...
      "DateBounds": {
        "properties": {
          "config": {
            "title": "Config",
            "type": "string"
          },
          "granularity": {
            "enum": [
              "day",
              "week",
              "month"
            ],
            "title": "Granularity",
            "type": "string"
          },
          "max": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Max"
          },
          "min": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Min"
          }
        },
        "required": [
          "config",
          "granularity",
          "min",
          "max"
        ],
        "title": "DateBounds",
        "type": "object"
      },
...
      "DeepdiveRequest": {
        "properties": {
          "clip": {
            "$ref": "#/components/schemas/ClipOpts"
          },
          "config": {
            "title": "Config",
            "type": "string"
          },
          "filter": {
            "$ref": "#/components/schemas/FilterOpts"
          },
          "granularity": {
            "default": "day",
            "enum": [
              "day",
              "week",
              "month"
            ],
            "title": "Granularity",
            "type": "string"
          },
          "group_values": {
            "additionalProperties": {
              "items": {
                "type": "string"
              },
              "type": "array"
            },
            "title": "Group Values",
            "type": "object"
          },
          "lifecycle_groups": {
            "anyOf": [
              {
                "items": {
                  "$ref": "#/components/schemas/LifecycleGroup"
                },
                "type": "array"
              },
              {
                "type": "null"
              }
            ],
            "title": "Lifecycle Groups"
          },
          "metrics": {
            "items": {
              "type": "string"
            },
            "minItems": 1,
            "title": "Metrics",
            "type": "array"
          },
          "mode": {
            "enum": [
              "histogram",
              "heatmap",
              "scatter"
            ],
            "title": "Mode",
            "type": "string"
          },
          "nbins": {
            "default": 50,
            "title": "Nbins",
            "type": "integer"
          },
          "normalize": {
            "default": false,
            "title": "Normalize",
            "type": "boolean"
          },
          "outliers_std": {
            "anyOf": [
              {
                "type": "number"
              },
              {
                "type": "null"
              }
            ],
            "title": "Outliers Std"
          },
          "panel": {
            "enum": [
              "derived",
              "user"
            ],
            "title": "Panel",
            "type": "string"
          },
          "range_groups": {
            "anyOf": [
              {
                "items": {
                  "$ref": "#/components/schemas/RangeGroup"
                },
                "type": "array"
              },
              {
                "type": "null"
              }
            ],
            "title": "Range Groups"
          },
          "ranges": {
            "items": {
              "$ref": "#/components/schemas/DateRange"
            },
            "minItems": 1,
            "title": "Ranges",
            "type": "array"
          }
        },
        "required": [
          "config",
          "panel",
          "mode",
          "metrics",
          "ranges"
        ],
        "title": "DeepdiveRequest",
        "type": "object"
      },
      "DeepdiveResponse": {
        "properties": {
          "config": {
            "title": "Config",
            "type": "string"
          },
          "granularity": {
            "enum": [
              "day",
              "week",
              "month"
            ],
            "title": "Granularity",
            "type": "string"
          },
          "heatmaps": {
            "items": {
              "$ref": "#/components/schemas/CorrMatrix"
            },
            "title": "Heatmaps",
            "type": "array"
          },
          "histograms": {
            "items": {
              "$ref": "#/components/schemas/HistogramSeries"
            },
            "title": "Histograms",
            "type": "array"
          },
          "missing": {
            "items": {
              "type": "string"
            },
            "title": "Missing",
            "type": "array"
          },
          "mode": {
            "enum": [
              "histogram",
              "heatmap",
              "scatter"
            ],
            "title": "Mode",
            "type": "string"
          },
          "panel": {
            "enum": [
              "derived",
              "user"
            ],
            "title": "Panel",
            "type": "string"
          },
          "scatters": {
            "items": {
              "$ref": "#/components/schemas/ScatterSeries"
            },
            "title": "Scatters",
            "type": "array"
          }
        },
        "required": [
          "config",
          "granularity",
          "panel",
          "mode",
          "histograms",
          "heatmaps",
          "scatters",
          "missing"
        ],
        "title": "DeepdiveResponse",
        "type": "object"
      },
...
      "HTTPValidationError": {
        "properties": {
          "detail": {
            "items": {
              "$ref": "#/components/schemas/ValidationError"
            },
            "title": "Detail",
            "type": "array"
          }
        },
        "title": "HTTPValidationError",
        "type": "object"
      },
...
    }
  },
  "info": {
    "title": "dashboard_api",
    "version": "0.1.0"
  },
  "openapi": "3.1.0",
  "paths": {
    "/api/data/config/{config_id}": {
      "get": {
        "description": "API endpoint: GET /api/data/config/{config_id} \u2014 resolved config for one game.\n\nDrives which tabs, metric groups, and date granularities the dashboard\nshows. This is also the snapshot the agent reads to decide which config to\nload and which tab to open for a natural-language data request.",
        "operationId": "get_config_api_data_config__config_id__get",
        "parameters": [
          {
            "in": "path",
            "name": "config_id",
            "required": true,
            "schema": {
              "title": "Config Id",
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ConfigDetail"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Get Config",
        "tags": [
          "data"
        ]
      }
    },
...
    "/api/data/date-bounds": {
      "get": {
        "description": "API endpoint: GET /api/data/date-bounds \u2014 min/max date in the source data.\n\nThe Stats-by-Date tab seeds its default window (last 30 days) from ``max``.",
        "operationId": "get_date_bounds_api_data_date_bounds_get",
        "parameters": [
          {
            "in": "query",
            "name": "config",
            "required": true,
            "schema": {
              "title": "Config",
              "type": "string"
            }
          },
          {
            "in": "query",
            "name": "granularity",
            "required": false,
            "schema": {
              "default": "day",
              "enum": [
                "day",
                "week",
                "month"
              ],
              "title": "Granularity",
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/DateBounds"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Get Date Bounds",
        "tags": [
          "data"
        ]
      }
    },
    "/api/data/deepdive": {
      "post": {
        "description": "API endpoint: POST /api/data/deepdive \u2014 Deep Dive tab.\n\nPre-binned histograms or a Pearson correlation matrix over the selected\nmetrics, per (cohort \u00d7 date-range), for the derived or user-level panel.",
        "operationId": "post_deepdive_api_data_deepdive_post",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/DeepdiveRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/DeepdiveResponse"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Post Deepdive",
        "tags": [
          "data"
        ]
      }
    },
...
    "/api/data/group-distribution": {
      "post": {
        "description": "API endpoint: POST /api/data/group-distribution \u2014 Stats-by-Group tab.\n\nPer (cohort \u00d7 date-range), the metric's distribution as a 5-number summary +\nmean/bootstrap-CI, so the frontend renders a box plot or a bar chart.",
        "operationId": "post_group_distribution_api_data_group_distribution_post",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/GroupDistributionRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/GroupDistributionResponse"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Post Group Distribution",
        "tags": [
          "data"
        ]
      }
    },
    "/api/data/group-values": {
      "get": {
        "description": "API endpoint: GET /api/data/group-values \u2014 distinct cohort values per\nuser-group column, for the Stats-by-Date cohort pickers.\n\n``values`` is the full (cached) vocabulary so labels never disappear when\nthe date range moves; with ``start``/``end`` the response also carries\n``available`` \u2014 the subset present in that range \u2014 which the pickers use\nto gray out values with no data in view.\n\nReturns ``{}`` when the config defines no user_group columns. Reads the same\nuser-level parquet as ``/series``.",
        "operationId": "get_group_values_api_data_group_values_get",
        "parameters": [
          {
            "in": "query",
            "name": "config",
            "required": true,
            "schema": {
              "title": "Config",
              "type": "string"
            }
          },
          {
            "in": "query",
            "name": "granularity",
            "required": false,
            "schema": {
              "default": "day",
              "enum": [
                "day",
                "week",
                "month"
              ],
              "title": "Granularity",
              "type": "string"
            }
          },
          {
            "in": "query",
            "name": "start",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Start"
            }
          },
          {
            "in": "query",
            "name": "end",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "End"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/GroupValues"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Get Group Values",
        "tags": [
          "data"
        ]
      }
    },
    "/api/data/series": {
      "post": {
        "description": "API endpoint: POST /api/data/series \u2014 metric time-series for a chart.\n\nReturns date-aligned series the frontend encodes into an ECharts chart.\nMetrics not present in the source parquet come back in ``missing`` rather\nthan failing the request, so the UI can flag them.",
        "operationId": "post_series_api_data_series_post",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/SeriesRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/SeriesResponse"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Post Series",
        "tags": [
          "data"
        ]
      }
    },
    "/api/data/summary-table": {
      "post": {
        "description": "API endpoint: POST /api/data/summary-table \u2014 Summary-table tab.\n\nOne grid for all metrics at once: rows are metrics (grouped by the config's\nmetric groups), columns are the selected cohort \u00d7 date-range combinations.\nEach cell is a metric's mean/median/quartiles for that column; the frontend\nflags one column as the reference and renders the rest as ``value (\u00b1%)``.\nWith ``pvalues=true`` each row also carries a Welch t-test (2 columns) or\none-way ANOVA (3+) p-value.",
        "operationId": "post_summary_table_api_data_summary_table_post",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/SummaryTableRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/SummaryTableResponse"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Post Summary Table",
        "tags": [
          "data"
        ]
      }
    },
...
  }
}